Exclude GWT library from gerrit_plugin_gwtui artifact
GWT user and dev libraries are shaded into gerrit_plugin_gwtui and
interferes the compilation of GWT plugins in standalone build mode.

The right way to fix that would be to avoid the inclusion of the
GWT libraries in gerrit_plugin_gwtui artifact in the first place.
As quick and dirty workaround strip the GWT libraries during fetch
of the plugin API from Central.

Change-Id: I76762cc1adf101f865ee58ff479a078a73211657