Automate sonar analysis for bazel building
The previous sonar analysis could be triggered for buck building only.
This change adds the sonar analysis implementation to bazlets so that
the sonar code quality measurement process can be automated for bazel
building.

Note that this implementation contains bazel plugin building, the
sonar-project.properties file generation and sonar-scanner triggering.
The test reporting is not included.

Change-Id: I7af492c60e562f85a9ec2810c853f70a8c4cd3ad