Automate sonar analysis for bazel building
The previous sonar analysis could be triggered for buck building only.
This change adds the sonar analysis implementation to bazlets so that
the sonar code quality measurement process can be automated for bazel
building.

This implementation contains bazel plugin building, the auto-generation
of sonar-project.properties file and sonar-scanner triggering. The test
reporting is not included.

A BUILD file is added for sonar which enables Gerrit plugins to run
sonar analysis by calling the sonar executable file from remote bazlets

With this BUILD file, a Gerrit standalone plugin could run the sonar by
following steps:
  1. Modify the SHA1 of the load_bazlets in the WORKSPACE file to be the
one of this commit once this change is merged.
  2. Create a sonar.sh file and add this command to it:
  `bazel query @com_googlesource_gerrit_bazlets//tools/sonar:sonar
--output location | sed s/BUILD:.*//`sonar.py .
  3. Run this sonar.sh

Change-Id: I7af492c60e562f85a9ec2810c853f70a8c4cd3ad