gerrit_plugin: Improve stamping error message
Bazel stamp error message is very misleading and hard to track down
and understand. A number of prerequisites must be met for the stamping
machinery to work properly.

Bazel's missing feature to dynamically construct MANIFEST.MF file and
pass such file to java_binary rule contributes to this problem: [1].

To rectify, split the stamp information retrieval in its own rule. The
reason for the failure is because bazel is using "set -o pipefail" for
genrules. This particular option sets the exit code of a pipeline to
that of the rightmost command to exit with a non-zero status, or to
zero if all commands of the pipeline exit successfully.

Given the self descriptive rule name: <plugin>__gen_stamp_info it is
clear where to look for, in tools/workspace-status.sh command.

Test Plan:

1. Apply this CL in bazlets repository
2. Clone reviewers plugin and switch to consume bazlets from local
directory
3. Patch tools/workspace-status.sh file and replace this line:
echo STABLE_BUILD_REVIEWERS_LABEL $(rev .)
with:
echo STABLE_BUILD_REVIEW_LABEL $(rev .)
4. Run bazel build :reviewers and notice that correct failing rule is
now reported:

  $ bazel build reviewers
  INFO: Analysed target //:reviewers (0 packages loaded).
  INFO: Found 1 target...
  ERROR: /home/davido/projects/reviewers/BUILD:3:1: Executing genrule //:reviewers__gen_stamp_info failed (Exit 1)
  Target //:reviewers failed to build

[1] https://github.com/bazelbuild/bazel/issues/2009

Change-Id: Ia2ad7238f77cafb89fa822b78c5d658ccedc8d11
(cherry picked from commit 4459b9706a6eedb8453a98d4434fb3bc4db84211)