Update patch set 1
Patch Set 1: Code-Review+1

The build of delete-project is still failing with a bunch of JSC errors.  I had run into those before, but couldn't find a way to suppress them.

http://paste.openstack.org/show/752396/

I'm not sure if it needs to be fixed in this commit, in a follow-up here, or on the delete-project itself.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1