RFC: Modify gerrit_api to allow it to be used for multiple API versions
Versions are defined in a dict that maps the version string to a list
of sha1s for the gerrit-plugin-api, gerrit-acceptance-framework and
gerrit-plugin-gwtui artifacts.

Defining multiple API versions in this way will allow us to switch to
a single branch of the bazlets project that can be used in plugins
across all branches. The plugin then selects the API version that it
wants to use when invoking gerrit_api.

This will reduce the amount of work we need to to do make bazlets
updates available to plugins. We only need to do it once, on the
common branch, and then update it on the plugins' earliest supported
branch. It then gets to the later plugin branches by merge.  This is
as opposed to the current process where we have to make a separate
bazlets update per plugin branch.

Change-Id: I28816b7d31927d7b9be52bd80486712b42cd85d4