Support bazelisk or bazel in tools/eclipse/project.py
It is recommended to use bazelisk instead of bazel directly, however
it is also perfectly reasonable to use bazel directly. Update the
eclipse script to default to looking for bazelisk then bazel, but to
continue to allow explicit overrides. This should allow people who
do not have bazelisk to not be broken, but people who have installed
bazelisk to have it be the program run by default.

This also causes whichever of the two was found to to go into
.bazel_path which will influence what eclipse uses.

Based on the change from Monty Taylor [1] in core Gerrit.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/234727

Bug: Issue 11359
Change-Id: I2490e479d16bdad64c771a38e6cce73decb062a6