package.bzl: Allow to pass pom file per artifact
This change extends the maven_package() rule to accept pom file per
artifact. This is required for medium to big size projects, where
pom file is specific per Bazel package.

The corresponding Bazel idiom is to export the pom file per Bazel
package using exports_files() rule and pass the label to the
maven_package() rule.

Say, in Bazel package foo, the package specific pom file is exported:

  exports_files(["pom.foo.xml"])

And in the tools/maven/BUILD file, the exported pom file is used and
passed in to the publishing machinery:

  maven_package(
      [...]
      pom = {
          "pom": "//foo:pom.foo.xml",
      [...]

Also change group to be an optional parameter in maven_package() rule
because with the pom file, group id is retrieved from the pom file.

Change-Id: I35d5cabf13d4a320ed53765bfdbae2f59d28436f