Bazel: Update time attribute of file entries in plugin artifact
Due to this Bazel issue: [1] non .class file entries are normalized in
singlejar Bazel utility. As the consequence even JavaScript and HTML
files have the timestamp: 2010-01-01 00:00. This makes it problematic
to refresh the plugin artifacts in browser because of the cache if a
new plugin version is deployed in Gerrit server.

To rectify, touch the content of the plugin artifact in stamp action.

The 'touch' command uses the system default timezone if none is
supplied. This makes zip files dependent on the timezone of the
generating machine.

Zip by default stores extended attributes (eg. ctime, atime) on Unix
systems. These can be further sources of non-determinism, so switch it
off with --no-extra.

[1] https://github.com/bazelbuild/bazel/issues/10789

Bug: Issue 12349
Change-Id: I5b3bbba8a3df15f37ab1a21c48768697e8bd2f36