Add repository rule to provide gerrit api version
In I0a786a91147 Gerrit-ApiVersion attribute was added to the MANIFEST.MF
in plugin rule in tree build mode. This change extends standalone build
mode to do the same.

There is one complication, though, because standalone build mode can use
release plugin API or SNAPSHOT and the resulting api version should be
switched as well. Use repository rule, that generates the version.txt
file and depend on it in plugin rule.

Test Plan:

1. Switch reviewers plugin to using bazlets repository from this commit
2. Confirm that building reviewers plugin with release plugin API would
add the following line to the MANIFEST.MF file:

  Gerrit-ApiVersion: 2.16.21

3. Change to gerrit workspace and publish SNAPSHOT plugin artifacts to
   the local Maven repository
4. Confirm that building reviewers plugin with SNAPSHOT plugin API would
add the following line to the MANIFEST.MF file:

  Gerrit-ApiVersion: 2.16.22-SNAPSHOT

Change-Id: I6f17e6c6bbfd4d6fba7bcda8a3f5fc1403c4dc4e