Update patch set 3
Patch Set 3:

This deduping is no longer needed, as the follow-up change that brought API version
to plugin jars has been abandoned.

Not sure if this change is still needed/wanted. And it will have a merge conflict with
the changes of I6f17e6c6bbfd4d6fba7bcda8a3f5fc1403c4dc4e.
But I find it nice to not have to comment/uncomment `load` lines when switching between
release and snapshot builds.

So I leave it around for now, so it can be discussed if de-duping is wanted.
Feel free to abandon otherwise.

Patch-set: 3