Update patch set 2
Patch Set 2:

I see, we still have this repository in gerrit core, but it stale and unsued since we switched to consuming JGit as submodule. So that we should probably clean up and remove this repository in gerrit core.

For the same reason, I wonder why would you need this in plugin? In both build modes, in tree and standalone, you would get a newer JGit releases, from the plugin API, because it is produced using JGit built as submodule that could be newer then any official release on Central or other Maven repositories.

Patch-set: 2