download to GERRIT_CACHE_HOME when set
When building a plugin as user `nobody`, download_file is unable to
write the artifacts to the cache since the user does not have a home
directory.  For a CI build I also need to be set the cache directory to
a predetermined value.

Introduce the optional `GERRIT_CACHE_HOME` environment variable to
support relocating downloaded artifacts.  When the environment variable
is not set, behavior is unchanged.

Update documentation to mention support for `GERRIT_CACHE_HOME`. While at
it, explain how to override the bazel repository and disk cache, would
have same a bit of time the first time I had to tweak them.

This was done in Gerrit with Ie4fac83928527e0e71b159b9500983234c2261ac

Change-Id: I0c4b3992d36cf40196f64251900f07e527d8a3c4