Refactor eclipse/project.py to make it testable
Move all the logic of tools/eclipse/project.py so it can be tested and
add a few tests.

Add a py_test target so one can run the tests from bazel using:

  bazel test --test_output=streamed //tools/eclipse:test

(I had to add gerrit_polymer() to WORKSPACE)

Add a bunch of tests and a helper to generate a coverage report.

Change-Id: I03582d70a5ec28117ab6d3b4c098d84e12e3d8dc