Migrate deprecated JavaInfo API usages
This is a purely mechanical change that replaces:

* transitive_runtime_deps with transitive_runtime_jars

The old and new field is simply alias, and there is no change in
semantics. This change will allow dropping the deprecated field from
JavaInfo, reducing bazel memory consumption. The field was marked
deprecated in Nov 2021, so can be safely dropped.

This change is needed to bump Bazel version to new release 7.0.

Change-Id: If22d6e870bcb22d3377dc10b223b8e8635db6834