Migrate maven_jar rule to bzlmod
Currently, the maven_jar rule is duplicated between the gerrit-project
and bazlets. To avoid that and to be better able to make use of bazel
modules the bazlets project should be used for both the gerrit project
and plugins.

This change makes the maven_jar rule a module extension that can be
used by other projects.

Change-Id: I311c980e5bf970e7b5f90405da3ea61028c129d0